refactor: innerclass로 정의해둔 RowMapper를 분리 및 패키지 이동
