+++ src/main/java/database/mapper/RowMapper.java
@@ -1,4 +1,4 @@
-package database.dao;
+package database.mapper;
 
 import java.sql.ResultSet;
 import java.sql.SQLException;
